+++ android/src/ti/admob/ViewProxy.java
@@ -1,6 +1,7 @@
 /**
  * Copyright (c) 2011 by Studio Classics. All Rights Reserved.
- * Author: Brian Kurzius
+ * Copyright (c) 2017-present by Axway Appcelerator. All Rights Reserved.
+ * Author: Brian Kurzius, Axway Appcelerator
  * Licensed under the terms of the Apache Public License
  * Please see the LICENSE included with this distribution for details.
  */
@@ -17,11 +18,13 @@ import org.appcelerator.titanium.view.TiUIView;
 import android.app.Activity;
 
 @Kroll.proxy(creatableInModule = AdmobModule.class)
-public class ViewProxy extends TiViewProxy implements OnLifecycleEvent {
+public class ViewProxy extends TiViewProxy implements OnLifecycleEvent
+{
 	private View adMob;
 	private static final String TAG = "AdMobViewProxy";
 
-	public ViewProxy() {
+	public ViewProxy()
+	{
 		super();
 	}
 
@@ -36,52 +39,60 @@ public class ViewProxy extends TiViewProxy implements OnLifecycleEvent {
 	}
 
 	@Override
-	protected KrollDict getLangConversionTable() {
+	protected KrollDict getLangConversionTable()
+	{
 		KrollDict table = new KrollDict();
 		table.put("title", "titleid");
 		return table;
 	}
 
 	@Override
-	public TiUIView createView(Activity activity) {
+	public TiUIView createView(Activity activity)
+	{
 		adMob = new View(this);
-		((TiBaseActivity)activity).addOnLifecycleEventListener(this);
+		((TiBaseActivity) activity).addOnLifecycleEventListener(this);
 		return adMob;
 	}
 
 	@Kroll.method
-	public void requestAd() {
+	public void requestAd(@Kroll.argument(optional = true) KrollDict parameters)
+	{
 		Log.d(TAG, "requestAd()");
-		adMob.requestAd();
+		adMob.requestAd(parameters);
 	}
 
 	@Kroll.method
-	public void requestTestAd() {
+	public void requestTestAd()
+	{
 		Log.d(TAG, "requestTestAd(): ");
 		adMob.requestTestAd();
 	}
 
 	@Override
-	public void onDestroy(Activity activity) {
+	public void onDestroy(Activity activity)
+	{
 		adMob.destroy();
 	}
 
 	@Override
-	public void onPause(Activity activity) {
+	public void onPause(Activity activity)
+	{
 		adMob.pause();
 	}
 
 	@Override
-	public void onResume(Activity activity) {
+	public void onResume(Activity activity)
+	{
 		adMob.resume();
 	}
 
 	@Override
-	public void onStart(Activity activity) {
+	public void onStart(Activity activity)
+	{
 	}
 
 	@Override
-	public void onStop(Activity activity) {
+	public void onStop(Activity activity)
+	{
 	}
-
 }