+++ ios/documentation/changelog.md
@@ -1,6 +1,8 @@
 # Change Log
 <pre>
-v2.2.0 Update Admob iOS SDK to 7.27.0, remove iAd adapter due to Google removal
+v2.3.0  [MOD-2196] Update AdMob to 7.31.0, conform to GDPR
+
+v2.2.0  Update Admob iOS SDK to 7.27.0, remove iAd adapter due to Google removal
 
 v2.1.0 	[MOD-2196] Support the iAd adapter